recruiter_bot/requirements.txt
python-telegram-bot==20.1
python-dotenv==1.0.0
pymongo==4.3.3